+++ test_umbra/test_illumina/test_alignment.py
@@ -7,7 +7,7 @@ subdirectory of an Illumina run directory on disk.
 
 from unittest.mock import Mock
 from pathlib import Path
-from umbra.illumina.run import Alignment
+from umbra.illumina.alignment import Alignment
 from ..test_common import TestBase
 
 class TestAlignment(TestBase):